refactor(test): 테스트 코드 리팩토링
__test__ 폴더 내 모든 파일을 test suite로 인식하여,
testUtils.js를 삭제하고, 각 테스트 파일에 이식하였습니다.
